single dir implementation + fixes
