--- .rustfmt.toml
@@ -1,1 +1,0 @@
-license_template_path = "LICENSE_TEMPLATE"
