+++ Cargo.toml
@@ -9,8 +9,7 @@ repository = "https://github.com/futureversecom/trn-trnnut-rs"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.6.9", default-features = false, features = ["derive"] }
-serde = { version = "1.0.145", default-features = false, features = ["derive"] }
-serde_json = { version = "1.0.64", default-features = false }
+serde = { version = "1.0.145", default-features = false, features = ["derive", "alloc"] }
 
 # pact = { git = "https://github.com/futureversecom/trn-pact", branch = "feat/wasm-pack", default-features = false }
 pact = { path = "../trn-pact/", default-features = false }
@@ -21,5 +20,4 @@ std = [
   "codec/std",
   "pact/std",
   "serde/std",
-  "serde_json/std"
 ]
