+++ js/Cargo.toml
@@ -11,14 +11,14 @@ repository = "https://github.com/futureversecom/trn-trnnut-rs"
 crate-type = ["cdylib"]
 
 [dependencies]
-trnnut-rs = { path = "../" }
+trnnut-rs = { path = "../", default-features = false }
+pact = { branch = "feat/wasm-pack", git = "https://github.com/futureversecom/trn-pact", default-features = false }
+
+codec = { package = "parity-scale-codec", version = "3.6.9", default-features = false, features = ["derive"] }
+serde-wasm-bindgen = "0.6.3"
 wasm-bindgen = { version = "0.2.63", features = ["serde-serialize"]  }
 clear_on_drop = { version = "0.2.3", features = ["no_cc"] }
-parity-scale-codec = { version = "3.6.9", features = ["derive"] }
-serde = { version = "1.0.145", features = ["derive"] }
 console_error_panic_hook = "0.1"
-
-pact = { branch = "feat/wasm-pack", git = "https://github.com/futureversecom/trn-pact", default-features = false }
 
 [profile.release]
 # disable loop vectorization for smaller wasm
