+++ src/lib.rs
@@ -5,28 +5,47 @@
 //! Delegated authority nut for TRN
 //!
 
-#![cfg_attr(not(feature = "std"), no_std)]
 #![warn(clippy::pedantic)]
 #![allow(clippy::module_name_repetitions, clippy::must_use_candidate)]
 
-#[cfg(not(feature = "std"))]
 extern crate alloc;
 
-#[cfg(feature = "std")]
-extern crate std as alloc;
-
+use alloc::fmt::{self, Display, Formatter};
 use codec::Input;
 pub use core::convert::TryFrom;
 
-mod trnnut;
-mod validation;
+pub mod trnnut;
+pub mod validation;
+pub mod method;
+pub mod module;
 
 pub use crate::{
-    trnnut::{v0, v0::TRNNutV0, RuntimeDomain, TRNNut},
+    trnnut::TRNNutV0,
     validation::ValidationErr,
 };
 
-mod test;
+#[cfg(test)]
+mod tests;
+
+pub const WILDCARD: &str = "*";
+
+/// A TRN module permission domain
+#[derive(Debug, Eq, PartialEq)]
+pub enum RuntimeDomain {
+    Method,
+    MethodArguments,
+    Module,
+}
+
+impl Display for RuntimeDomain {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Method => write!(f, "method"),
+            Self::MethodArguments => write!(f, "method arguments"),
+            Self::Module => write!(f, "module"),
+        }
+    }
+}
 
 pub trait PartialDecode: Sized {
     /// decode an input which is not including the version as the up front two bytes
