+++ src/method.rs
@@ -5,21 +5,20 @@
 //! Delegated method permissions of TRNNut for use in TRN
 //!
 
-use alloc::string::ToString;
-use alloc::vec::Vec;
+use alloc::{
+    string::{String, ToString},
+    vec::Vec,
+};
 use codec::{Decode, Encode, Input, Output};
 use core::convert::TryFrom;
 use pact::types::Contract as PactContract;
-#[cfg(feature = "std")]
-use serde::{Deserialize, Serialize};
 
 const BLOCK_COOLDOWN_MASK: u8 = 0x01;
 const CONSTRAINTS_MASK: u8 = 0x02;
 const MAX_CONSTRAINTS: usize = 256;
 
 /// A TRN permission domain module method
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[cfg_attr(test, derive(Clone, Debug, Eq, PartialEq))]
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
 pub struct Method {
     pub name: String,
     pub block_cooldown: Option<u32>,
