+++ src/trnnut.rs
@@ -5,24 +5,15 @@
 //! Version 0 TRNNut type.
 //!
 
-#[cfg(test)]
-mod tests;
+
 
 use alloc::vec::Vec;
 use codec::{Decode, Encode, Input, Output};
 use core::convert::TryFrom;
 use pact::{interpreter::interpret, types::PactType};
 
-pub mod method;
-pub mod module;
-
-use super::RuntimeDomain;
-use crate::{PartialDecode, ValidationErr};
+use crate::{module, PartialDecode, RuntimeDomain, ValidationErr, WILDCARD};
 use module::Module;
-#[cfg(feature = "std")]
-use serde::{Deserialize, Serialize};
-
-use super::WILDCARD;
 
 pub const MAX_MODULES: usize = 256;
 pub const MAX_METHODS: usize = 128;
@@ -30,8 +21,7 @@ pub const VERSION_BYTES: [u8; 2] = [0, 0];
 pub const MAX_TRNNUT_BYTES: usize = u16::max_value() as usize;
 
 /// A TRN permission domain struct for embedding in doughnuts
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[cfg_attr(test, derive(Clone, Debug, Eq, PartialEq))]
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
 pub struct TRNNutV0 {
     pub modules: Vec<Module>,
 }
